+++ Sources/ContaServiceArqui/Models/ContaCorrente.swift
@@ -16,9 +16,9 @@ final class ContaCorrente: Model, Content, ProtocoloContaCorrente, @unchecked Se
     // MARK: - Propriedades Fluent (substituem as private(set) anteriores)
     @ID(key: .id)              var id: UUID?
     @Field(key: "nome")        var nome: String
-    @Field(key: "saldo")       var saldo: Decimal
-    @Field(key: "salario_atual")    var salarioAtual: Decimal
-    @Field(key: "salario_anterior") var salarioAnterior: Decimal
+    @Field(key: "saldo") var saldo: Double
+    @Field(key: "salario_atual") var salarioAtual: Double
+    @Field(key: "salario_anterior") var salarioAnterior: Double
 
     var negativado: Bool { saldo < 0 }
 
@@ -34,14 +34,14 @@ final class ContaCorrente: Model, Content, ProtocoloContaCorrente, @unchecked Se
     }
 
     // MARK: - Lógica de domínio (igual ao monolito)
-    public func registraNovoSalario(valor: Decimal) -> Resultado {
+    public func registraNovoSalario(valor: Double) -> Resultado {
         self.salarioAnterior = self.salarioAtual
         self.salarioAtual = valor
         print("Seu novo salario é de: \(valor)")
         return .sucesso(novoValor: salarioAtual)
     }
 
-    public func depositar(valor: Decimal) -> Resultado {
+    public func depositar(valor: Double) -> Resultado {
         guard valor > 0 else {
             return .falha(erro: "Valor de depósito inválido. Deve ser maior que zero.")
         }
@@ -50,7 +50,7 @@ final class ContaCorrente: Model, Content, ProtocoloContaCorrente, @unchecked Se
         return .sucesso(novoValor: saldo)
     }
 
-    public func sacar(valor: Decimal) -> Resultado {
+    public func sacar(valor: Double) -> Resultado {
         saldo -= valor
         print("Saque de R$ \(valor) realizado. Novo saldo: R$ \(saldo)")
         return .sucesso(novoValor: saldo)
